フロント側で作成したAPIを使用して適切にコンポーネント分割を適応
