+++ app/post/types/PostType.ts
@@ -5,3 +5,9 @@ export type PostType = {
   authorId: number;
   createdAt: string;
 };
+export type Posts = {
+  id: number;
+  title: string;
+  content: string;
+  createdAt: string;
+};
